+++ packages/vibrant-components/src/lib/SelectField/index.ts
@@ -1,1 +1,2 @@
 export { SelectField } from './SelectField';
+export type { SelectFieldProps } from './SelectFieldProps';