+++ packages/vibrant-components/src/lib/TextField/TextField.tsx
@@ -21,6 +21,8 @@ export const TextField = withTextFieldVariation(
     autoComplete = 'none',
     onValueChange,
     clearable = false,
+    onFocus,
+    onBlur,
     ...restProps
   }) => {
     const inputRef = useRef<TextInputRef>(null);
@@ -66,8 +68,16 @@ export const TextField = withTextFieldVariation(
             disabled={disabled}
             autoCapitalize={autoCapitalize}
             autoComplete={autoComplete}
-            onFocus={() => setIsFocused(true)}
-            onBlur={() => setIsFocused(false)}
+            onFocus={() => {
+              onFocus?.();
+
+              setIsFocused(true);
+            }}
+            onBlur={() => {
+              onBlur?.();
+
+              setIsFocused(false);
+            }}
             onValueChange={({ value }) => {
               setValue(value);