+++ packages/vibrant-core/src/lib/Box/BoxProps.ts
@@ -108,6 +108,12 @@ export type BoxElements =
   | 'svg'
   | 'ul';
 
+type ComponentProps<Component extends ComponentType | undefined> = Component extends undefined
+  ? undefined
+  : Component extends ComponentType<infer ComponentProps>
+  ? ComponentProps
+  : undefined;
+
 export type BoxProps<
   BaseComponent extends ComponentType | undefined = undefined,
   ElementName extends BoxElements | undefined = undefined
@@ -117,21 +123,21 @@ export type BoxProps<
   id?: string;
   onLayout?: (layoutEvent: LayoutEvent) => void;
 } & DistributiveOmit<
-    BaseComponent extends ComponentType<infer BaseComponentProps>
-      ? BaseComponentProps
-      : (ElementName extends keyof JSX.IntrinsicElements
+    ComponentProps<BaseComponent> extends undefined
+      ? (ElementName extends keyof JSX.IntrinsicElements
           ? JSX.IntrinsicElements[ElementName]
           : Record<never, never>) & {
           ref?: Ref<HTMLElement>;
           children?: ReactElementChild | ReactElementChild[];
-        },
+        }
+      : ComponentProps<BaseComponent>,
     keyof SystemProps
   >;
 
 export type { LayoutEvent };
 
 export const interpolation = injectContext(
-  createInterpolation(systemProps, { display: 'flex', boxSizing: 'border-box' })
+  createInterpolation(systemProps, { display: 'flex', boxSizing: 'border-box', position: 'relative' })
 );
 
 export const shouldForwardProp = createShouldForwardProp(systemPropNames);