+++ packages/vibrant-core/src/lib/StackedPortalProvider/StackedPortalProviderProps.ts
@@ -1,0 +1,48 @@
+import type { ReactElementChild } from '../../types';
+
+export type Position = 'bottom' | 'top';
+
+export type Offset = number | string;
+export type RenderedIndex = number;
+
+export type StackOptions = {
+  position: Position;
+  id: string;
+  order: number;
+  offset?: Offset;
+  safeAreaInset?: boolean;
+};
+
+export type OffsetChangeEvent = (_: { offset?: Offset; renderedIndex?: RenderedIndex }) => void;
+
+export type StackedPortalData = {
+  [position in Position]: {
+    [id in string]: {
+      [order in string]: {
+        height: number;
+        offset: Offset;
+        safeAreaInset: boolean;
+        onOffsetChange?: OffsetChangeEvent;
+      };
+    };
+  };
+};
+
+export type RegisterPortal = (_: StackOptions & { onOffsetChange: OffsetChangeEvent }) => void;
+export type UnregisterPortal = (_: StackOptions) => void;
+export type ChangePortalHeight = (_: StackOptions & { height: number }) => void;
+export type GetCalculatedOffset = (_: StackOptions) => { offset?: Offset; renderedIndex?: RenderedIndex };
+
+export type StackedPortalContextValue = {
+  registerPortal: RegisterPortal;
+  unregisterPortal: UnregisterPortal;
+  changePortalHeight: ChangePortalHeight;
+  getCalculatedOffset: GetCalculatedOffset;
+};
+
+export type StackedPortalProviderProps = {
+  children: ReactElementChild;
+  priorityOrder: {
+    [key in Position]: string[];
+  };
+};